- renamed Components; - global styles without module.css; - file structure refactoring;
