+++ react-draw-sim/src/App.module.css
@@ -8,4 +8,97 @@
   -ms-flex-direction: column;
   flex-direction: column;
 }
+
+.main {
+  -webkit-box-flex: 1;
+  -ms-flex: 1 0 auto;
+  flex: 1 0 auto;
+  padding: 3em;
+}
+
+@media (max-width: 767px) {
+  .main {
+    padding: 2em;
+  }
+}
+
+@media (max-width: 424px) {
+  .main {
+    padding: 1em;
+  }
+}
+
+.main .errors {
+  opacity: 0;
+  line-height: 2;
+  font-size: 14px;
+  font-family: 'Courier New', Courier, monospace;
+  color: #E9E9E9;
+}
+
+.main .lgErrors {
+  display: block;
+  text-align: center;
+}
+
+@media (max-width: 1023px) {
+  .main .lgErrors {
+    display: none;
+  }
+}
+
+.main .lgErrors.listErrors {
+  opacity: 1;
+  text-align: left;
+}
+
+.main .lgErrors.arrListErrors {
+  opacity: 1;
+}
+
+.main .lgErrors.conditionErrors {
+  opacity: 1;
+  text-align: right;
+}
+
+.main .smErrors {
+  display: none;
+  margin-bottom: 0.5em;
+  text-align: center;
+}
+
+@media (max-width: 1023px) {
+  .main .smErrors {
+    display: block;
+  }
+}
+
+.main .smErrors.listErrors, .main .smErrors.arrListErrors, .main .smErrors.conditionErrors {
+  opacity: 1;
+}
+
+.main .btn {
+  margin-top: 3em;
+  padding: 1em;
+  border-radius: 1em;
+  text-transform: uppercase;
+  outline: none;
+  border: none;
+  color: #E9E9E9;
+  background-color: #4EB1BA;
+  font-weight: bold;
+  -webkit-transition-duration: 0.6s;
+  transition-duration: 0.6s;
+  cursor: pointer;
+}
+
+.main .btn:hover {
+  background-color: #E9E9E9;
+  color: #00103f;
+}
+
+.main .btn:active {
+  background-color: #808080;
+  color: #00103f;
+}
 /*# sourceMappingURL=App.module.css.map */
