+++ react-draw-sim/src/Components/Data/Data.scss
