+++ react-draw-sim/src/Components/Data/DataCountInput/DataCountInput.js
@@ -1,5 +1,5 @@
 import { React } from "react";
-import "../DataContent.css";
+import "../Data.css";
 
 const DataCountInput = (props) => {
   const onChangeValue = (e) => {
