+++ react-draw-sim/src/Components/Teams/Teams.js
@@ -1,5 +1,5 @@
 import { React } from 'react';
-import Team from './Team';
+import Team from './Team/Team';
 import "./Teams.css";
 
 const Teams = (props) => {
