+++ react-draw-sim/src/Components/Teams/Teams.scss
